wordcloud-cli-for-japanese/pyproject.toml
[tool.poetry]
name = "wordcloud-cli-for-japanese"
version = "0.9.0"
description = ""
authors = ["Niten Nashiki <n.nashiki.work@gmail.com>"]
include = ["src/wordcloud_cli_for_japanese/data/*"]

[tool.poetry.dependencies]
python = "^3.8"
wordcloud = "^1.8.1"
mecab-python3 = "^1.0.3"
unidic-lite = "^1.0.7"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
wordcloud_cli_for_japanese = "wordcloud_cli_for_japanese.core:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
